Winner, Delete y Loser Tests seaparados y new UpdateUserNameTest
