+++ tests/Feature/UserRollDiceControllerTest.php
@@ -13,7 +13,7 @@ use App\Models\DiceRoll;
 
 use function PHPUnit\Framework\assertCount;
 
-class RollDiceControllerTest extends TestCase
+class UserRollDiceControllerTest extends TestCase
 {
 
     use RefreshDatabase;
@@ -39,33 +39,4 @@ class RollDiceControllerTest extends TestCase
         $response->dump();
         
     }
-
-    public function test_user_can_delete_rolls(){
-
-        Role::create(['name' => 'user']);
-        $user = User::factory()->create()->assignRole('user');
-
-        DiceRoll::create([ // create a roll
-            'user_id' => $user->id,
-            'dice1' => 1,
-            'dice2' => 1,
-            'total' => 2,
-            'isWin' => false
-        ]);
-
-        $this->assertCount(1, DiceRoll::all()); // check if there is 1 roll
-
-        $response = $this->actingAs($user, 'api')->json('DELETE', route('user.api.v1.delete', ['id' => $user->id]))
-        ->assertStatus(200); //
-
-        $this->assertCount(0, DiceRoll::all()); // check if there is 0 roll
-
-        $this->assertDatabaseMissing('dice_rolls', [
-            'user_id' => $user->id
-        ]);
-
-        
-
-    }
-
 }
